refactor, general Analyst agent interface
