+++ listen-kit/src/twitter/user_info.rs
@@ -1,5 +1,4 @@
-use super::{ApiResponse, TwitterApi};
-use crate::data::twitter::TwitterApiError;
+use super::{ApiResponse, TwitterApi, TwitterApiError};
 use anyhow::Result;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
